requirements.txt
Flask==2.3.2
torch==2.0.1  # Replace with the specific version you have installed
Pillow==10.0.0
opencv-python==4.8.0.74
